new ape testing utils module
